fix wrong eventbus and wrong event
