+++ src/main/java/vazkii/patchouli/common/base/Patchouli.java
@@ -1,10 +1,9 @@
 package vazkii.patchouli.common.base;
 
-import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.common.MinecraftForge;
-import net.minecraftforge.fml.DistExecutor;
+import net.minecraftforge.event.OnDatapackSyncEvent;
+import net.minecraftforge.event.RegisterCommandsEvent;
 import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import net.minecraftforge.fml.loading.FMLEnvironment;
@@ -12,9 +11,12 @@ import net.minecraftforge.fml.loading.FMLEnvironment;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-import vazkii.patchouli.client.base.ClientProxy;
+import vazkii.patchouli.common.book.BookRegistry;
+import vazkii.patchouli.common.command.OpenBookCommand;
 import vazkii.patchouli.common.handler.LecternEventHandler;
 import vazkii.patchouli.common.item.PatchouliItems;
+import vazkii.patchouli.common.network.NetworkHandler;
+import vazkii.patchouli.common.network.message.MessageReloadBookContents;
 
 @Mod(Patchouli.MOD_ID)
 public class Patchouli {
@@ -31,7 +33,7 @@ public class Patchouli {
 		var modEventBus = FMLJavaModLoadingContext.get().getModEventBus();
 		modEventBus.addListener(this::onCommonSetup);
 		MinecraftForge.EVENT_BUS.addListener(this::onRegisterCommands);
-		MinecraftForge.EVENT_BUS.addListener(this::onServerStart);
+		MinecraftForge.EVENT_BUS.addListener(this::onDatapackSync);
 		MinecraftForge.EVENT_BUS.addListener(LecternEventHandler::onRightClick);
 		PatchouliItems.init();
 	}
@@ -47,10 +49,7 @@ public class Patchouli {
 		OpenBookCommand.register(e.getDispatcher());
 	}
 
-	private void onServerStart(FMLServerStartedEvent evt) {
-		MinecraftServer server = evt.getServer();
-		// Also reload contents when someone types /reload
-		ResourceManagerReloadListener listener = m -> MessageReloadBookContents.sendToAll(server);
-		((ReloadableResourceManager) server.getResourceManager()).registerReloadListener(listener);
+	private void onDatapackSync(OnDatapackSyncEvent evt) {
+		MessageReloadBookContents.send(evt.getPlayer());
 	}
 }
